refactor: merge replaceVariables helpers
