+++ packages/api-client/src/helpers/replaceVariables.test.ts
@@ -3,11 +3,27 @@ import { describe, expect, it } from 'vitest'
 import { replaceVariables } from './replaceVariables'
 
 describe('replaceVariables', () => {
-  it('replaces variables', async () => {
+  it('replaces variables with a single bracket', () => {
     expect(
-      replaceVariables('http://{baseUrl}/foobar', {
-        baseUrl: 'example.com',
+      replaceVariables('foo{bar}foo', {
+        bar: 'foo',
       }),
-    ).toBe('http://example.com/foobar')
+    ).toMatchObject('foofoofoo')
+  })
+
+  it('replaces variables with a double bracket', () => {
+    expect(
+      replaceVariables('foo{{bar}}foo', {
+        bar: 'foo',
+      }),
+    ).toMatchObject('foofoofoo')
+  })
+
+  it('replaces variables with a custom callback', () => {
+    expect(
+      replaceVariables('foo{{bar}}foo', (match) =>
+        match === 'bar' ? `<span>foo</span>` : '',
+      ),
+    ).toMatchObject('foo<span>foo</span>foo')
   })
 })
