--- packages/api-reference/src/helpers/replaceVariables.test.ts
@@ -1,29 +1,0 @@
-import { describe, expect, it } from 'vitest'
-
-import { replaceVariables } from './replaceVariables'
-
-describe('replaceVariables', () => {
-  it('replaces variables with a single bracket', () => {
-    expect(
-      replaceVariables('foo{bar}foo', {
-        bar: 'foo',
-      }),
-    ).toMatchObject('foofoofoo')
-  })
-
-  it('replaces variables with a double bracket', () => {
-    expect(
-      replaceVariables('foo{{bar}}foo', {
-        bar: 'foo',
-      }),
-    ).toMatchObject('foofoofoo')
-  })
-
-  it('replaces variables with a custom callback', () => {
-    expect(
-      replaceVariables('foo{{bar}}foo', (match) =>
-        match === 'bar' ? `<span>foo</span>` : '',
-      ),
-    ).toMatchObject('foo<span>foo</span>foo')
-  })
-})
